refactor: share export download helper
