+++ src/lib/export-utils.ts
@@ -4,6 +4,17 @@ type ExportFileNameOptions = {
   timestamp?: Date;
 };
 
+export function downloadBlob(blob: Blob, fileName: string): void {
+  const url = URL.createObjectURL(blob);
+  const link = document.createElement("a");
+  link.href = url;
+  link.download = fileName;
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+  URL.revokeObjectURL(url);
+}
+
 function formatTimestamp(timestamp: Date): string {
   return timestamp
     .toISOString()
