+++ src/lib/midi-export.ts
@@ -55,26 +55,3 @@ export function exportMidi(options: MidiExportOptions): Uint8Array {
   // Convert to Uint8Array
   return midi.toArray();
 }
-
-/**
- * Download MIDI file to the user's computer
- * @param midiData - Uint8Array containing MIDI file data
- * @param fileName - Desired file name
- */
-export function downloadMidiFile(midiData: Uint8Array, fileName: string): void {
-  // Create a blob from the byte array
-  // Cast to any to avoid TypeScript issues with ArrayBufferLike vs ArrayBuffer
-  const blob = new Blob([midiData as any], { type: "audio/midi" });
-
-  // Create a download link and trigger it
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement("a");
-  link.href = url;
-  link.download = fileName;
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-
-  // Clean up the URL
-  URL.revokeObjectURL(url);
-}
